requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4

cryptography==2.3.1
dj-database-url==0.5.0

django-filter==2.0.0
django-heroku==0.3.1
django-paypal==0.2.5
django-tables==0.2
django-tables2==2.0.6
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.7
Markdown==3.0.1
Pillow==5.3.0
pipenv==2018.11.26
psycopg2==2.7.6.1
psycopg2-binary==2.8.3
pycosat==0.6.3
pycparser==2.18
pyOpenSSL==18.0.0
PySocks==1.6.8


requests==2.20.0
ruamel-yaml==0.15.46
six==1.11.0
South==1.0.2
urllib3==1.24.2
virtualenv==16.2.0
virtualenv-clone==0.4.0
whitenoise==4.1.2
win-inet-pton==1.0.1
wincertstore==0.2
